+++ CHANGELOG.md
@@ -4,3 +4,7 @@ Initial release.
 ## Added
 * GZipCompressionDelegatingHandler
 * DeflateCompressionDelegatingHandler
+* HttpStatusCodeExtensions
+  - 422 Unprocessable Entity
+  - 424 Failed Dependency
+  - 429 Too Many Requests